refactor: WIP initial refactor
